+++ HISTORY.md
@@ -3,6 +3,7 @@
 * Modified `stats_dump` to return internal stats as string, allowing for output in Jupyter notebooks [#403](https://github.com/TileDB-Inc/TileDB-Py/pull/403)
 * `open_dataframe` now uses the underlying Array's `nonempty_domain` [#409](https://github.com/TileDB-Inc/TileDB-Py/pull/409)
 * Added `__repr__` to `Array` and `Ctx` [#413](https://github.com/TileDB-Inc/TileDB-Py/pull/413)
+* `tiledb.open` now supports `timestamp` keyword argument [#419](https://github.com/TileDB-Inc/TileDB-Py/pull/419)
 
 # TileDB-Py 0.7.0 Release Notes